Fix some compiler warnings
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/3222



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@14234 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/common/propname.cpp
@@ -196,7 +196,7 @@ U_CAPI int32_t U_EXPORT2
 u_getPropertyValueEnum(UProperty property,
                        const char* alias) {
     return load() ? PNAME->getPropertyValueEnum(property, alias)
-                  : UCHAR_INVALID_CODE;
+                  : (int32_t)UCHAR_INVALID_CODE;
 }
 
 /* data swapping ------------------------------------------------------------ */
